Restructure files and change names of the aplication, new name AIDB, add models, etc...
