+++ src-tauri/tauri.conf.json
@@ -200,7 +200,7 @@
         "fullscreen": false,
         "height": 600,
         "resizable": false,
-        "title": "Android-IOS AIDB Commands",
+        "title": "AIDB Commands",
         "width": 800
       }
     ]
