+++ src/models/device.model.ts
@@ -1,0 +1,7 @@
+import { DeviceTypeEnum } from "./enums/device.enum";
+
+export interface Device {
+    id: string;
+    name: string;
+    type: DeviceTypeEnum
+}
