+++ src/models/enums/device.enum.ts
@@ -1,0 +1,6 @@
+export enum DeviceTypeEnum {
+    NONE = 0,
+    IPHONE = 1,
+    ANDROID = 2,
+    OTHER = 3,
+}
