+++ src/views/configurationScreen/components/InputLocation.tsx
